Move video/video tests to the new folder structure
