--- src/Context/Video/Module/Video/Test/PhpUnit/VideoModuleFunctionalTestCase.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace CodelyTv\Context\Video\Module\Video\Test\PhpUnit;
-
-use CodelyTv\Context\Video\Module\Video\Domain\VideoRepository;
-use CodelyTv\Test\Context\Video\PhpUnit\VideoContextFunctionalTestCase;
-
-abstract class VideoModuleFunctionalTestCase extends VideoContextFunctionalTestCase
-{
-
-}
