+++ tests/Context/Video/Module/User/Domain/ScalaVideoCreatedDomainEventStub.php
@@ -7,7 +7,7 @@ namespace CodelyTv\Test\Context\Video\Module\User\Domain;
 use CodelyTv\Context\Video\Module\User\Domain\ScalaVideoCreatedDomainEvent;
 use CodelyTv\Context\Video\Module\User\Domain\UserId;
 use CodelyTv\Context\Video\Module\Video\Domain\VideoId;
-use CodelyTv\Context\Video\Module\Video\Test\Stub\VideoIdStub;
+use CodelyTv\Test\Context\Video\Module\Video\Application\Create\VideoIdStub;
 
 final class ScalaVideoCreatedDomainEventStub
 {
