+++ tests/Context/Video/Module/Video/Application/Find/FindVideoQueryStub.php
@@ -2,10 +2,11 @@
 
 declare(strict_types=1);
 
-namespace CodelyTv\Context\Video\Module\Video\Test\Stub;
+namespace CodelyTv\Test\Context\Video\Module\Video\Application\Find;
 
 use CodelyTv\Context\Video\Module\Video\Application\Find\FindVideoQuery;
 use CodelyTv\Context\Video\Module\Video\Domain\VideoId;
+use CodelyTv\Test\Context\Video\Module\Video\Domain\VideoIdStub;
 
 final class FindVideoQueryStub
 {
