+++ tests/Context/Video/Module/Video/Domain/VideoUrlStub.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace CodelyTv\Context\Video\Module\Video\Test\Stub;
+namespace CodelyTv\Test\Context\Video\Module\Video\Domain;
 
 use CodelyTv\Context\Video\Module\Video\Domain\VideoUrl;
 use CodelyTv\Test\Stub\UrlStub;
