+++ tests/Context/Video/Module/Video/VideoModuleFunctionalTestCase.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace CodelyTv\Test\Context\Video\Module\Video;
+
+use CodelyTv\Test\Context\Video\VideoContextFunctionalTestCase;
+
+abstract class VideoModuleFunctionalTestCase extends VideoContextFunctionalTestCase
+{
+}
