+++ tests/Context/Video/Module/VideoComment/Application/Publish/PublishVideoCommentCommandStub.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CodelyTv\Context\Video\Module\Video\Test\Stub;
+namespace CodelyTv\Test\Context\Video\Module\Video\Application\Create;
 
 use CodelyTv\Context\Video\Module\Video\Domain\VideoId;
 use CodelyTv\Context\Video\Module\VideoComment\Contract\PublishVideoCommentCommand;
